Add two rest services with tests and documentation
One service that allows to register a user and the other one to get details about him. Moreover, interceptor allows to log information about http request and response. The readme file contains installation prerequisites, commands line to run the app, use case diagram and a schema of the application architecture.